backend/api-gateway/requirements.txt
# backend/api-gateway/requirements.txt
# Dépendances Python pour le microservice API Gateway

# Framework web asynchrone
fastapi~=0.100.0

# Serveur ASGI pour FastAPI
uvicorn[standard]~=0.22.0

# Client HTTP asynchrone pour les appels inter-services
httpx~=0.27.0

# Validation des données (dépendance de FastAPI)
pydantic~=2.0.0
pydantic-settings~=2.0.0 # Pour la gestion des paramètres de configuration

# Hashing des mots de passe
passlib[bcrypt]~=1.7.0

# Gestion des jetons JWT
python-jose[cryptography]~=3.3.0

# Logging (si des bibliothèques de logging spécifiques sont utilisées, sinon la lib standard suffit)
# Par exemple: python-json-logger ou loguru si nécessaire pour un formatage spécifique

# Dépendances partagées (ne sont pas installées directement via requirements.txt ici,
# mais sont copiées dans l'image Docker et importées comme modules locaux)
# shared (config, models, exceptions, utils)
